backend/requirements.txt
# backend/requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4

# Database (PostgreSQL support for production)
psycopg2-binary==2.9.9

# AI/ML packages - Updated versions
torch==2.8.0
torchvision==0.19.0
transformers==4.46.0
huggingface-hub==0.26.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.11.4
matplotlib==3.8.4
seaborn==0.13.2

# Geospatial packages (for future PostGIS support)
# GDAL==3.8.4
# geopandas==0.14.4
# rasterio==1.3.10

# Web3/Blockchain (for NFT features)
web3==6.20.0
eth-account==0.11.0

# Development and utilities
python-decouple==3.8
pillow==10.3.0
requests==2.31.0

# Testing
pytest==8.1.1
pytest-django==4.8.0